moved export bin generation to utils
